+++ pkg/webhook/validating/webhook_test.go
@@ -7,14 +7,12 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	admissionv1beta1 "k8s.io/api/admission/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	mmetrics "github.com/slok/kubewebhook/mocks/observability/metrics"
 	"github.com/slok/kubewebhook/pkg/log"
 	"github.com/slok/kubewebhook/pkg/observability/metrics"
 	"github.com/slok/kubewebhook/pkg/webhook/validating"
@@ -22,6 +20,10 @@ import (
 
 func getPodJSON() []byte {
 	pod := &corev1.Pod{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "v1",
+			Kind:       "Pod",
+		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "testPod",
 			Namespace: "testNS",
@@ -42,14 +44,14 @@ func getFakeValidator(valid bool, message string) validating.Validator {
 }
 
 func TestPodAdmissionReviewValidation(t *testing.T) {
-	tests := []struct {
-		name        string
+	tests := map[string]struct {
+		cfg         validating.WebhookConfig
 		validator   validating.Validator
 		review      *admissionv1beta1.AdmissionReview
 		expResponse *admissionv1beta1.AdmissionResponse
 	}{
-		{
-			name:      "A review of a Pod with a valid validator result should return allowed.",
+		"A static webhook review of a Pod with a valid validator result should return allowed.": {
+			cfg:       validating.WebhookConfig{Name: "test", Obj: &corev1.Pod{}},
 			validator: getFakeValidator(true, "valid test chain"),
 			review: &admissionv1beta1.AdmissionReview{
 				Request: &admissionv1beta1.AdmissionRequest{
@@ -68,8 +70,50 @@ func TestPodAdmissionReviewValidation(t *testing.T) {
 				},
 			},
 		},
-		{
-			name:      "A review of a Pod with a invalid validator result should return not allowed.",
+
+		"A static webhook review of a Pod with a invalid validator result should return not allowed.": {
+			cfg:       validating.WebhookConfig{Name: "test", Obj: &corev1.Pod{}},
+			validator: getFakeValidator(false, "invalid test chain"),
+			review: &admissionv1beta1.AdmissionReview{
+				Request: &admissionv1beta1.AdmissionRequest{
+					UID: "test",
+					Object: runtime.RawExtension{
+						Raw: getPodJSON(),
+					},
+				},
+			},
+			expResponse: &admissionv1beta1.AdmissionResponse{
+				UID:     "test",
+				Allowed: false,
+				Result: &metav1.Status{
+					Message: "invalid test chain",
+				},
+			},
+		},
+
+		"A dynamic webhook review of a Pod with a valid validator result should return allowed.": {
+			cfg:       validating.WebhookConfig{Name: "test"},
+			validator: getFakeValidator(true, "valid test chain"),
+			review: &admissionv1beta1.AdmissionReview{
+				Request: &admissionv1beta1.AdmissionRequest{
+					UID: "test",
+					Object: runtime.RawExtension{
+						Raw: getPodJSON(),
+					},
+				},
+			},
+			expResponse: &admissionv1beta1.AdmissionResponse{
+				UID:     "test",
+				Allowed: true,
+				Result: &metav1.Status{
+					Status:  metav1.StatusSuccess,
+					Message: "valid test chain",
+				},
+			},
+		},
+
+		"A dynamic webhook review of a Pod with a invalid validator result should return not allowed.": {
+			cfg:       validating.WebhookConfig{Name: "test"},
 			validator: getFakeValidator(false, "invalid test chain"),
 			review: &admissionv1beta1.AdmissionReview{
 				Request: &admissionv1beta1.AdmissionRequest{
@@ -89,27 +133,16 @@ func TestPodAdmissionReviewValidation(t *testing.T) {
 		},
 	}
 
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
 			assert := assert.New(t)
 			require := require.New(t)
 
-			// Mocks.
-			mm := &mmetrics.Recorder{}
-			mm.On("IncAdmissionReview", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Once()
-			mm.On("ObserveAdmissionReviewDuration", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Once()
-
-			cfg := validating.WebhookConfig{
-				Name: "test",
-				Obj:  &corev1.Pod{},
-			}
-
-			wh, err := validating.NewWebhook(cfg, test.validator, nil, mm, log.Dummy)
+			wh, err := validating.NewWebhook(test.cfg, test.validator, nil, nil, log.Dummy)
 			require.NoError(err)
 			gotResponse := wh.Review(context.TODO(), test.review)
 
 			assert.Equal(test.expResponse, gotResponse)
-			mm.AssertExpectations(t)
 		})
 	}
 }